pyproject.toml
[project]
name = "fyp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.2.1",
]
